+++ src/libs/AxelarQueryAPI.ts
@@ -15,8 +15,9 @@ import {
 import { throwIfInvalidChainIds } from "../utils";
 import { loadChains } from "../chains";
 import s3 from "./TransactionRecoveryApi/constants/s3";
-import { BigNumber } from "ethers";
+import { BigNumber, BigNumberish } from "ethers";
 import { ChainInfo } from "src/chains/types";
+import { BigNumberUtils } from "./BigNumberUtils";
 
 interface TranslatedTransferRateLimitResponse {
   incoming: string;
@@ -37,7 +38,7 @@ export interface AxelarQueryAPIFeeResponse {
   executionFee: string;
   executionFeeWithMultiplier: string;
   gasMultiplier: number;
-  gasLimit: number;
+  gasLimit: BigNumberish;
   minGasPrice: string;
   apiResponse: any;
   isExpressSupported: boolean;
@@ -98,10 +99,12 @@ export class AxelarQueryAPI {
   public async getEVMEvent(sourceChainId: string, srcTxHash: string, srcEventId: number) {
     await throwIfInvalidChainIds([sourceChainId], this.environment);
     await this.initQueryClientIfNeeded();
-    return this.axelarQueryClient.evm.Event({
-      chain: sourceChainId,
-      eventId: `${srcTxHash}-${srcEventId}`,
-    });
+    return this.axelarQueryClient.evm
+      .Event({
+        chain: sourceChainId,
+        eventId: `${srcTxHash}-${srcEventId}`,
+      })
+      .catch(() => undefined);
   }
 
   public async getConfirmationHeight(chain: string) {
@@ -251,7 +254,7 @@ export class AxelarQueryAPI {
     sourceChainId: EvmChain | string,
     destinationChainId: EvmChain | string,
     sourceChainTokenSymbol: GasToken | string,
-    gasLimit: number = DEFAULT_ESTIMATED_GAS,
+    gasLimit: BigNumberish = DEFAULT_ESTIMATED_GAS,
     gasMultiplier = 1.1,
     minGasPrice = "0",
     gmpParams?: GMPParams
@@ -276,14 +279,16 @@ export class AxelarQueryAPI {
       throw new Error("Failed to estimate gas fee");
     }
 
-    const destGasFeeWei = parseUnits(
-      (gasLimit * Number(destToken.gas_price)).toFixed(destToken.decimals),
+    const destGasFeeWei = BigNumberUtils.multiplyToGetWei(
+      BigNumber.from(gasLimit),
+      destToken.gas_price,
       destToken.decimals
     );
-
     const minDestGasFeeWei = BigNumber.from(gasLimit).mul(minGasPrice); //minGasPrice already provided by the user in wei
-    const srcGasFeeWei = parseUnits(
-      (gasLimit * Number(sourceToken.gas_price)).toFixed(sourceToken.decimals),
+
+    const srcGasFeeWei = BigNumberUtils.multiplyToGetWei(
+      BigNumber.from(gasLimit),
+      sourceToken.gas_price,
       sourceToken.decimals
     );