+++ src/libs/TransactionRecoveryApi/constants/chain/mainnet.ts
@@ -15,8 +15,10 @@ export const rpcMap: Record<EvmChain, string> = {
   [EvmChain.KAVA]: "https://evm.kava.io",
   [EvmChain.FILECOIN]: "https://rpc.ankr.com/filecoin",
   [EvmChain.OPTIMISM]: "https://optimism-mainnet.public.blastapi.io",
-  [EvmChain.BASE]: "",
-  [EvmChain.LINEA]: "",
+  [EvmChain.BASE]: "https://developer-access-mainnet.base.org",
+  [EvmChain.LINEA]: "https://rpc.linea.build",
+  [EvmChain.POLYGON_ZKEVM]: "https://zkevm.polygonscan.com",
+  [EvmChain.MANTLE]: "https://rpc.mantle.xyz",
 };
 
 export const networkInfo: Record<EvmChain, Network> = {
@@ -25,7 +27,7 @@ export const networkInfo: Record<EvmChain, Network> = {
     name: EvmChain.FANTOM,
   },
   [EvmChain.BASE]: {
-    chainId: -1, // does not exist yet
+    chainId: 8453,
     name: EvmChain.BASE,
   },
   [EvmChain.FILECOIN]: {
@@ -77,7 +79,15 @@ export const networkInfo: Record<EvmChain, Network> = {
     name: EvmChain.OPTIMISM,
   },
   [EvmChain.LINEA]: {
-    chainId: -1, // does not exist yet
+    chainId: 59144,
     name: EvmChain.LINEA,
+  },
+  [EvmChain.POLYGON_ZKEVM]: {
+    chainId: 1101,
+    name: EvmChain.POLYGON_ZKEVM,
+  },
+  [EvmChain.MANTLE]: {
+    chainId: 5000,
+    name: EvmChain.MANTLE,
   },
 };