+++ src/libs/types/index.ts
@@ -16,6 +16,7 @@ export enum EvmChain {
   AVALANCHE = "avalanche",
   FANTOM = "fantom",
   POLYGON = "polygon",
+  POLYGON_ZKEVM = "polygon-zkevm",
   MOONBEAM = "moonbeam",
   AURORA = "aurora",
   BINANCE = "binance",
@@ -27,6 +28,7 @@ export enum EvmChain {
   FILECOIN = "filecoin",
   OPTIMISM = "optimism",
   LINEA = "linea",
+  MANTLE = "mantle",
 }
 
 export enum CosmosChain {
@@ -151,6 +153,8 @@ export enum GasToken {
   FILECOIN = "FIL",
   OSMO = "OSMO",
   AXL = "AXL",
+  POLYGON_ZKEVM = "ETH",
+  MANTLE = "MNT",
 }
 
 export interface AddGasOptions {
@@ -158,6 +162,8 @@ export interface AddGasOptions {
   refundAddress?: string;
   estimatedGasUsed?: number;
   gasMultipler?: number;
+  eventIndex?: number;
+  destChain?: string;
   evmWalletDetails?: EvmWalletDetails;
 }
 
@@ -267,6 +273,7 @@ export const isNativeToken = (chain: string, selectedToken: GasToken): boolean =
     kava: GasToken.KAVA,
     base: GasToken.BASE,
     filecoin: GasToken.FILECOIN,
+    linea: GasToken.ETH,
   };
   return nativeTokenMap[chain]?.toLowerCase() === selectedToken?.toLowerCase();
 };